+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.1] 2026-04-06
+
+### Deprecated
+
+* `Order::big` in favor of `Order::Big`
+* `Order::little` in favor of `Order::Little`
+* `Order::network` in favor of `Order::Network`
+
 ## [0.3.0] 2025-11-25
 
 ### Changed